chapter-5/anpr/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-storage
google-cloud-pubsub
pytesseract
imutils
opencv-python-headless
matplotlib
numpy
